move helper functions to pocketbase.ts
